Improve reader and writer testing
Both the reader and writer were restrictive in adding extra tests due
to their existing implementation. By converting these tests into test
tables, more tests can be added (included failure states).

It was necessary to convert both the reader and writer tests at the
same time as they were sharing global variables (which have been
removed).

The history variables contained metafied characters which made it
difficult to handle as a Go string and required combining with other
encoding styles. As part of this improvement, golden files were used
for the history file as this allowed it to be shared between the reader
and writer tests as well as solve the issues with encoding.