Add support for standard history format
The ZSH standard history format is the default history format. It is
necessary to enable extended history format as part of the ZSH
configuration.

Both formats can be mixed within the same history file. Identification
of lines using the extended format can be found by the prefix of `: `
at the start of each command.

This change adds support for the standard history format. When reading,
time and elapsed are set to `0` as this information is not available.
When writing, if the time is set to `0`, a standard format line will be
written which will only contain the command.

Unit tests for both the reader and writer have been added.